#70 css 수정
