+++ client/src/components/common/useFilter.jsx
@@ -26,14 +26,12 @@ export const useFilter = (initialFilters = {}) => {
   };
 };
 
-// 직무 필터 옵션
 export const JOB_OPTIONS = [
   { value: "직군·직무", label: "직군·직무" },
   { value: "프론트엔드 개발자", label: "프론트엔드 개발자" },
   { value: "백엔드 개발자", label: "백엔드 개발자" },
 ];
 
-// 질문유형 필터 옵션
 export const TYPE_OPTIONS = [
   { value: "질문유형", label: "질문유형" },
   { value: "인성", label: "인성" },
