+++ client/src/pages/myPage/_components/bookmark/settings/constants.js
@@ -1,13 +1,13 @@
-// 북마크 관련 상수
+export const PAGE_SIZE = 4;
 
-// 페이지당 표시할 아이템 수
-export const PAGE_SIZE = 3;
-
-// 그리드 레이아웃 스타일
 export const GRID_TEMPLATE = "34px 118px 50px 1fr 80px";
 
-// 텍스트 색상 설정
 export const TEXT_COLORS = {
-  normal: "text-gray-500",
-  accent: "text-zik-main"
+  normal: "text-zik-text",
+  accent: "text-zik-main",
+  header: "text-xs font-semibold text-zik-text",
+  title: "text-zik-text",
+  description: "text-zik-text",
+  button: "text-white",
+  link: "text-zik-main"
 };
