+++ client/src/pages/myPage/_components/bookmark/settings/index.js
@@ -1,4 +1,3 @@
-// 북마크 설정 통합 내보내기
 export * from './constants';
 export * from './utils';
 export * from './components';
