+++ client/src/pages/myPage/_components/question/settings/utils.js
@@ -1,10 +1,7 @@
-// 데이터 변환 유틸리티
 import { faqData } from "@/data/faqData";
 import { ACTIONS } from './constants';
+import { SORT_OPTIONS } from "@/components/common/useFilter";
 
-/**
- * faqData를 면접 결과 데이터 형식으로 변환
- */
 export const convertToResultData = (faqItem) => ({
   id: faqItem.id,
   title: faqItem.career || "일반 직무",
@@ -17,24 +14,18 @@ export const convertToResultData = (faqItem) => ({
   feedback: `${faqItem.career || "일반 직무"} 포지션에 대한 답변은 전반적으로 좋았습니다. ${faqItem.recommendation || "기술적 이해도가 높고 실무 경험이 풍부해 보입니다. 다만 일부 질문에서 구체적인 사례 설명이 부족했습니다. 프로젝트 경험을 더 자세히 언급하면 좋을 것 같습니다."}`
 });
 
-/**
- * 초기 데이터 준비
- */
 export const prepareInitialData = () => 
   faqData
     .filter(item => item.type !== "일반")
     .map(convertToResultData);
 
-/**
- * 필터링 및 정렬 함수
- */
 export const filterAndSortResults = (results, filterType, starredItems) => {
   const sortedResults = [...results];
   
   switch (filterType) {
-    case "최신순":
+    case SORT_OPTIONS.LATEST:
       return sortedResults.sort((a, b) => new Date(b.date) - new Date(a.date));
-    case "북마크순":
+    case SORT_OPTIONS.BOOKMARK:
       return sortedResults.filter(item => starredItems.includes(String(item.id)));
     default:
       return sortedResults;
